requirements.txt
flask
moviepy
Pillow
